* Added a "Timeline" namespace * Moved Pinned* types to the Sdk namespace
