+++ AlembicImporter/Packages/com.unity.formats.alembic/Runtime/Scripts/Misc/PinnedArray.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
 
-namespace UnityEngine.Formats.Alembic.Importer
+namespace UnityEngine.Formats.Alembic.Sdk
 {
 
     public class PinnedObject<T> : IDisposable
@@ -102,10 +102,6 @@ namespace UnityEngine.Formats.Alembic.Importer
         public static implicit operator IntPtr(PinnedArray<T> v) { return v == null ? IntPtr.Zero : v.Pointer; }
     }
 
-}
-
-namespace UnityEngine.Formats.Alembic.Sdk
-{
     #region dirty
     public static class PinnedListImpl
     {
